First_Open_AI_Agents_SDK/pyproject.toml
[project]
name = "first-open-ai-agents-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
